fix: xp on screen

+++ cmd/canary/main.go
@@ -347,7 +347,7 @@ func run(o runOpts, log *slog.Logger) error {
 	world.OnPlayerStatsChange = func(p *game.Player) {
 		protocol.SendPlayerStats(p)
 	}
-	world.OnTextMessage = func(p *game.Player, class uint8, value uint32, text string) {
+	world.OnTextMessage = func(p *game.Player, class uint8, value uint64, text string) {
 		protocol.SendExpMessage(p, value, text)
 	}
 	world.OnPlayerDeath = func(p *game.Player, killer game.Creature) {
