+++ internal/game/world.go
@@ -48,7 +48,7 @@ type World struct {
 	OnPlayerStatsChange func(p *Player)
 
 	// OnTextMessage pushes a text message / animated text packet (0xB4) to a player.
-	OnTextMessage func(p *Player, class uint8, value uint32, text string)
+	OnTextMessage func(p *Player, class uint8, value uint64, text string)
 
 	// OnPlayerDeath is fired when a player's health reaches 0. The protocol
 	// layer applies the death penalty, teleports the player to their temple,
